Moved email address for local testing to env variables
