+++ dot-env.example
@@ -1,4 +1,7 @@
-# CONDUCTOR_URL is the URL of the Conductor app.
+# Default email address for passwordless sign in.
+VITE_DEFAULT_PASSWORDLESS_EMAIL_ADDRESS= # Set your email address for local development
+
+# URL of the Conductor app.
 VITE_CONDUCTOR_URL=http://localhost:5173/
 
 # Required Firebase config.
