--- packages/shared/src/lib/accounts.shared.ts
@@ -1,8 +1,0 @@
-import type {EmailAddress} from '@shared/types/emails.types';
-
-export function isInternalAccount(args: {readonly email: EmailAddress}): boolean {
-  const {email} = args;
-
-  // TODO: Move this to an environment variable.
-  return email === 'wenger.jacob@gmail.com';
-}
